chore: add transactionals, finals, custom exceptions, cleanup code in scheduling service

+++ SchedulingService/src/main/java/com/example/empsched/scheduling/service/OrganisationService.java
@@ -2,17 +2,21 @@ package com.example.empsched.scheduling.service;
 
 import com.example.empsched.scheduling.entity.Employee;
 import com.example.empsched.scheduling.entity.Organisation;
+import com.example.empsched.scheduling.exceptions.NotOrganisationOwnerException;
 import com.example.empsched.scheduling.repository.EmployeeRepository;
 import com.example.empsched.scheduling.repository.OrganisationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
+@Transactional(propagation = Propagation.REQUIRES_NEW)
 public class OrganisationService {
 
     private final OrganisationRepository organisationRepository;
@@ -27,7 +31,10 @@ public class OrganisationService {
         return savedOrganisation;
     }
 
-    public void deleteOrganisation(final UUID id) {
+    public void deleteOrganisation(final UUID id, final UUID requestingOrganisationId) {
+        if (!id.equals(requestingOrganisationId)) {
+            throw new NotOrganisationOwnerException();
+        }
         // TODO handle cascade delete or constraints (e.g., check for associated employees)
         organisationRepository.deleteById(id);
     }
